Change: cambios del frontend arreglados

+++ Front-End/src/index.js
@@ -2,20 +2,19 @@ const btn = document.querySelector('.btn')
 const lista = document.querySelector('.taskList')
 const title = document.querySelector('#titulo')
 const descripcion = document.querySelector('#descrip')
+const titleDelete = document.querySelector('#tituloDelete')
+
 const mostraNotas = async () => {
   const datos = await fetch('http://localhost:3001/notas')
   const data = await datos.json()
 
   data.forEach(data => {
     lista.innerHTML += `
-    <li class="bg-blue-100 w-3/4 m-auto border-black border-solid border-2 rounded-md p-2">
-      <div class="grid grid-cols-2 ">
-      <section class=" ml-5 w-3/4">
-          <h1>${data.content}</h1>
-          <p>${data.descript}</p>
-      </section>
-      <button class="btnDelete pt-1 pb-1 p-2 w-1/4 rounded-md border-solid border-black border-2 self-center hover justify-self-end bg-yellow-400">Modificar</button>
-      </div>
+    <li class="bg-white w-3/4 m-auto border-black border-solid border-2 rounded-md p-2" style = 'background-color: white'>
+        <section class=" ml-5 w-auto">
+            <h1 class = 'text-center'>${data.content}</h1>
+            <p style = 'margin-left: 40px'>${data.descript}</p>
+        </section>
     </li>
     `
   })
@@ -25,21 +24,23 @@ btn.addEventListener('click', async (e) => {
   e.preventDefault()
   const titulo = document.querySelector('.title').value
   const descrip = document.querySelector('.desc').value
-  const OPTIONS = {
-    method: 'post',
-    headers: {
-      'Content-type': 'application/json'
-    },
-    body: JSON.stringify({ titulo, descrip }),
-    mode: 'cors'
+  if (!titulo && !descrip) { alert('No colocaste nada') } else {
+    const OPTIONS = {
+      method: 'post',
+      headers: {
+        'Content-type': 'application/json'
+      },
+      body: JSON.stringify({ titulo, descrip }),
+      mode: 'cors'
+    }
+    fetch('http://localhost:3001/notas/nota/', OPTIONS)
+      .then(res => console.log(res.ok))
+      .catch(e => { console.log(e) })
+    lista.textContent = ''
+    mostraNotas()
+    title.value = ''
+    descripcion.value = ''
   }
-  fetch('http://localhost:3001/notas/nota/', OPTIONS)
-    .then(res => console.log(res.ok))
-    .catch(e => { console.log(e) })
-  lista.textContent = ''
-  mostraNotas()
-  title.value = ''
-  descripcion.value = ''
 })
 
 mostraNotas()
@@ -47,13 +48,16 @@ const btnDelete = document.querySelector('.btnDelete')
 
 btnDelete.addEventListener('click', async () => {
   const title = document.querySelector('.titleDelete').value
-  console.log(title)
-  await fetch(`http://localhost:3001/notas/nota/${title}`, {
-    method: 'delete',
-    mode: 'cors'
-  })
-  lista.textContent = ''
-  mostraNotas()
+  if (!title) { alert('No colocaste nada') } else {
+    console.log(title)
+    await fetch(`http://localhost:3001/notas/nota/${title}`, {
+      method: 'delete',
+      mode: 'cors'
+    })
+    lista.textContent = ''
+    titleDelete.value = ''
+    mostraNotas()
+  }
 })
 
 /*
